cAER: cameracalibration: refine and optimize results.
git-svn-id: https://svn.code.sf.net/p/jaer/code/cAER/trunk@8426 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ modules/cameracalibration/calibration.hpp
@@ -40,7 +40,7 @@ private:
 	Mat cameraMatrix;
 	Mat distCoeffs;
 
-	vector<Point2f> undistortEventMap;
+	vector<Point2i> undistortEventMap;
 	Mat undistortRemap1;
 	Mat undistortRemap2;
